requirements.txt
streamlit 
python-dotenv
google-generativeai
google-genai
gTTS
Pillow


#